requirements.txt
attrs==20.2.0
awsebcli==3.19.0
bcrypt==3.2.0
blessed==1.17.10
botocore==1.17.63
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1.1
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.16.0
idna==2.7
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
paramiko==2.7.2
pathspec==0.5.9
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.0
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
SQLAlchemy==1.3.19
termcolor==1.1.0
texttable==1.6.3
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.2.0
